restructured OsuLegacyDecoder into LegacyDecoder
Beatmap works, Storyboard not...
